# schemas/openning.go
package schemas

import (
	"time"

	"gorm.io/gorm"
)

type Openning struct {
	gorm.Model // unrap
	Role       string
	Company    string
	Location   string
	Remote     bool
	Link       string
	Salary     int64
}

type OpenningResponse struct {
	ID        uint      `json:"id"`
	CreatedAt time.Time `json:"created_at"`
	UpdatedAt time.Time `json:"updated_at"`
	DeletedAt time.Time `json:"deleted_at,omitempty"`
	Role      string    `json:"role"`
	Company   string    `json:"company"`
	Location  string    `json:"location"`
	Remote    bool      `json:"remote"`
	Link      string    `json:"link"`
	Salary    int64     `json:"salary"`
}